Add jetstream and horizon

+++ app/Console/Commands/ResetAppCommand.php
@@ -14,7 +14,7 @@ class ResetAppCommand extends Command
      *
      * @var string
      */
-    protected $signature = 'app:reset';
+    protected $signature = 'app:reset {--redis}';
 
     /**
      * The console command description.
@@ -31,9 +31,10 @@ class ResetAppCommand extends Command
     public function handle()
     {
         $this->task('Reset DB', fn() => Artisan::call('migrate:fresh --seed'));
-        $this->task('Reset Redis', fn() => RedisDB::client()->flushAll());
         $this->task('Reset Meilisearch', fn() => app(Client::class)->deleteAllIndexes());
-
+        if ($this->option('redis')) {
+            $this->task('Reset Redis', fn() => RedisDB::client()->flushAll());
+        }
         return 0;
     }
 }
